Split Root_Selection
